Update the Newtonsoft.Json parser to version 12.0.3
Also update the project to use the new `setup` command and .NET 5 tooling. Installing the [.NET 5 SDK](https://dotnet.microsoft.com/download/dotnet/5.0) is required for this parser to work.
