--- parsers/test_Json.NET/bin/Release/netcoreapp1.0/publish/test_Json.NET.deps.json
@@ -1,49 +1,0 @@
-{
-  "runtimeTarget": {
-    "name": ".NETCoreApp,Version=v1.0",
-    "signature": "09db60146a5b8a0d40c5ea0fb7485ab3bbdd4a1a"
-  },
-  "compilationOptions": {},
-  "targets": {
-    ".NETCoreApp,Version=v1.0": {
-      "test_Json.NET/1.0.0": {
-        "dependencies": {
-          "Newtonsoft.Json": "9.0.1"
-        },
-        "runtime": {
-          "test_Json.NET.dll": {}
-        }
-      },
-      "Newtonsoft.Json/9.0.1": {
-        "dependencies": {
-          "System.Runtime.Serialization.Primitives": "4.1.1"
-        },
-        "runtime": {
-          "lib/netstandard1.0/Newtonsoft.Json.dll": {}
-        }
-      },
-      "System.Runtime.Serialization.Primitives/4.1.1": {
-        "runtime": {
-          "lib/netstandard1.3/System.Runtime.Serialization.Primitives.dll": {}
-        }
-      }
-    }
-  },
-  "libraries": {
-    "test_Json.NET/1.0.0": {
-      "type": "project",
-      "serviceable": false,
-      "sha512": ""
-    },
-    "Newtonsoft.Json/9.0.1": {
-      "type": "package",
-      "serviceable": true,
-      "sha512": "sha512-U82mHQSKaIk+lpSVCbWYKNavmNH1i5xrExDEquU1i6I5pV6UMOqRnJRSlKO3cMPfcpp0RgDY+8jUXHdQ4IfXvw=="
-    },
-    "System.Runtime.Serialization.Primitives/4.1.1": {
-      "type": "package",
-      "serviceable": true,
-      "sha512": "sha512-HZ6Du5QrTG8MNJbf4e4qMO3JRAkIboGT5Fk804uZtg3Gq516S7hAqTm2UZKUHa7/6HUGdVy3AqMQKbns06G/cg=="
-    }
-  }
-}
