--- parsers/test_Json.NET/bin/Release/netcoreapp1.0/publish/test_Json.NET.runtimeconfig.json
@@ -1,8 +1,0 @@
-{
-  "runtimeOptions": {
-    "framework": {
-      "name": "Microsoft.NETCore.App",
-      "version": "1.0.1"
-    }
-  }
-}
