+++ parsers/test_dotnet_newtonsoft/.gitignore
@@ -1,0 +1,2 @@
+bin/
+obj/
